modern_autogen_v07/requirements.txt
autogen-agentchat
autogen-ext[openai,web-surfer,magentic-one,file-surfer]
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
